refactor dpm wave lifecycle actions
